+++ skore/tests/unit/reports/cross_validation/metrics/test_summarize.py
@@ -85,6 +85,7 @@ def test_binary_classification_forest(forest_binary_classification_data):
             "Brier score",
             "Fit time (s)",
             "Predict time (s)",
+            "Score",
         },
         expected_estimator_name="RandomForestClassifier",
     )
@@ -118,6 +119,7 @@ def test_binary_classification_svc(svc_binary_classification_data):
             "ROC AUC",
             "Fit time (s)",
             "Predict time (s)",
+            "Score",
         },
         expected_estimator_name="SVC",
     )
@@ -142,6 +144,7 @@ def test_multiclass_classification_forest(forest_multiclass_classification_data)
             "ROC AUC",
             "Predict time (s)",
             "Fit time (s)",
+            "Score",
         },
         expected_estimator_name="RandomForestClassifier",
     )
@@ -172,6 +175,7 @@ def test_multiclass_classification_svc(svc_multiclass_classification_data):
             "Recall",
             "Fit time (s)",
             "Predict time (s)",
+            "Score",
         },
         expected_estimator_name="SVC",
     )
@@ -198,6 +202,7 @@ def test_regression(linear_regression_data):
             "MAPE",
             "Fit time (s)",
             "Predict time (s)",
+            "Score",
         },
         expected_estimator_name="LinearRegression",
     )
@@ -221,6 +226,7 @@ def test_multioutput_regression(linear_regression_multioutput_data):
             "MAPE",
             "Fit time (s)",
             "Predict time (s)",
+            "Score",
         },
         expected_estimator_name="LinearRegression",
     )
@@ -246,6 +252,7 @@ def test_without_predict_proba(custom_classifier_no_predict_proba_data):
             "Recall",
             "Fit time (s)",
             "Predict time (s)",
+            "Score",
         },
         expected_estimator_name="CustomClassifierPredictOnly",
     )